Implement stat for all ports

+++ source/ports/x86_64-linux/sys/types.h
@@ -7,18 +7,18 @@
 extern "C" {
 #endif
 
-typedef int64_t dev_t;
-typedef int64_t ino_t;
-typedef int64_t mode_t;
-typedef int64_t nlink_t;
-typedef int64_t uid_t;
-typedef int64_t gid_t;
-typedef int64_t time_t;
-typedef int64_t blksize_t;
-typedef int64_t blkcnt_t;
-typedef int64_t pid_t;
-typedef int64_t off_t;
-typedef int64_t ssize_t;
+typedef uint64_t dev_t;
+typedef uint64_t ino_t;
+typedef uint32_t mode_t;
+typedef uint64_t nlink_t;
+typedef uint32_t uid_t;
+typedef uint32_t gid_t;
+typedef int64_t  time_t;
+typedef int64_t  blksize_t;
+typedef int64_t  blkcnt_t;
+typedef int32_t  pid_t;
+typedef int64_t  off_t;
+typedef int64_t  ssize_t;
 
 #ifdef __cplusplus
 }
